Remove DDL file and move MERGE into a separate file.
